+++ src/OneBlinkFormBase.tsx
@@ -50,6 +50,7 @@ import { OnUploadAttachmentContext } from './hooks/useOnUploadAttachment'
 import { injectOptionsAcrossAllElements } from './services/injectableOptions'
 import MaterialIcon from './components/MaterialIcon'
 import ReCAPTCHA from 'react-google-recaptcha'
+import ValidationErrorsCard from './components/ValidationErrorsCard'
 
 export type OneBlinkReadOnlyFormProps = {
   /**
@@ -601,14 +602,6 @@ function OneBlinkFormBase({
       }
       if (formElementsValidation) {
         console.log('Validation errors', formElementsValidation)
-        bulmaToast.toast({
-          message: 'Please fix validation errors',
-          type: 'is-danger',
-          extraClasses: 'ob-toast cypress-invalid-submit-attempt',
-          duration: 4000,
-          pauseOnHover: true,
-          closeOnClick: true,
-        })
         return
       }
       if (!checkBsbsCanBeSubmitted(submissionData.submission)) {
@@ -962,7 +955,9 @@ function OneBlinkFormBase({
               value={obFormContainerHTMLElementRef.current}
             >
               <div
-                className="ob-form-container"
+                className={clsx('ob-form-container', {
+                  'is-showing-pages': isShowingMultiplePages,
+                })}
                 ref={obFormContainerHTMLElementRef}
               >
                 <form
@@ -1429,6 +1424,13 @@ function OneBlinkFormBase({
                       </MaterialIcon>
                     </Modal>
                   </React.Fragment>
+                )}
+                {!!formElementsValidation && hasAttemptedSubmit && (
+                  <ValidationErrorsCard
+                    formElementsValidation={formElementsValidation}
+                    setPageId={setPageId}
+                    currentPage={currentPage}
+                  />
                 )}
               </div>
             </OneBlinkFormContainerContext.Provider>