+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Comprehensive validation errors toast
 - `FormElementFile` open files in a new tab
 
 ## [6.8.0] - 2024-10-08